Initial setup/config of frontend
