+++ app/src/App.js
@@ -5,14 +5,14 @@ import { LoadingContainer } from "drizzle-react-components";
 import "./App.css";
 
 import drizzleOptions from "./drizzleOptions";
-import MyContainer from "./MyContainer";
+import InterfaceContainer from "./components/InterfaceContainer";
 
 class App extends Component {
   render() {
     return (
       <DrizzleProvider options={drizzleOptions}>
         <LoadingContainer>
-          <MyContainer />
+          <InterfaceContainer />
         </LoadingContainer>
       </DrizzleProvider>
     );
