--- app/src/MyContainer.js
@@ -1,15 +1,0 @@
-import MyComponent from "./MyComponent";
-import { drizzleConnect } from "drizzle-react";
-
-const mapStateToProps = state => {
-  return {
-    accounts: state.accounts,
-    SimpleStorage: state.contracts.SimpleStorage,
-    TutorialToken: state.contracts.TutorialToken,
-    drizzleStatus: state.drizzleStatus,
-  };
-};
-
-const MyContainer = drizzleConnect(MyComponent, mapStateToProps);
-
-export default MyContainer;
