+++ app/src/drizzleOptions.js
@@ -1,6 +1,4 @@
-import SimpleStorage from "./contracts/SimpleStorage.json";
-import ComplexStorage from "./contracts/ComplexStorage.json";
-import TutorialToken from "./contracts/TutorialToken.json";
+import Bankshot from "./contracts/Bankshot.json";
 
 const options = {
   web3: {
@@ -10,9 +8,9 @@ const options = {
       url: "ws://127.0.0.1:9545",
     },
   },
-  contracts: [SimpleStorage, ComplexStorage, TutorialToken],
+  contracts: [Bankshot],
   events: {
-    SimpleStorage: ["StorageSet"],
+    
   },
   polls: {
     accounts: 1500,
